+++ .vscode/settings.json
@@ -9,7 +9,7 @@
     "**/.fvm": true
   },
   "editor.codeActionsOnSave": {
-    "source.fixAll": true,
-    "source.organizeImports": true
+    "source.fixAll": "explicit",
+    "source.organizeImports": "explicit"
   }
 }
