+++ front/lib/app/config/routes.dart
@@ -1,7 +1,7 @@
-import 'package:front/domain/presentation/home/home.dart';
-import 'package:front/domain/presentation/login/email_singin.dart';
-import 'package:front/domain/presentation/login/login.dart';
-import 'package:front/domain/presentation/login/signIn.dart';
+import 'package:front/app/domain/presentation/home/home.dart';
+import 'package:front/app/domain/presentation/login/email_singin.dart';
+import 'package:front/app/domain/presentation/login/login.dart';
+import 'package:front/app/domain/presentation/login/signIn.dart';
 
 class RouteName {
   static const home = '/';
