+++ front/lib/app/domain/presentation/login/component/checkbox_text_row.dart
